Remove nose from validation_efield and validation_rd
